refactor, dataset read error unresolved

+++ include/SAG.h
@@ -7,7 +7,7 @@ using namespace Eigen;
 
 void SAG_logistic(VectorXd &w, const MatrixXd &Xt, VectorXd &y, const MatrixXd &XtTest, \
 	 VectorXd &yTest, VectorXd &d, VectorXd &g, string filename, double lambda=0.1, double eta=0.1, \
-	int maxIter=60, int batchSize=1, int pass=20, double a=1, double b=1, int gamma=1,  int maxRunTime=60);
+	int maxIter=60, int batchSize=1, int pass=20, double a=1, double b=1, double gamma=1,  int maxRunTime=60);
 
 void SAG_logistic(VectorXd &w, const SparseMatrix<double> &Xt, int* innerIndices, int* outerStarts, const VectorXd &y, double lambda, double eta, VectorXd d, VectorXd g, \
 	int maxIter, int batchSize, int pass, double a, double b, int gamma, const MatrixXd &XtTest, \
