+++ include/SVRG.h
@@ -7,7 +7,7 @@ using namespace Eigen;
 
 void SVRG_logistic(VectorXd &w, const MatrixXd &Xt, VectorXd &y, const MatrixXd &XtTest, \
 	 VectorXd &yTest, VectorXd &wtilde, VectorXd &G, string filename, double lambda=0.1, double eta=0.1, \
-	int maxIter=60, int batchSize=1, int pass=20, double a=1, double b=1, int gamma=1,  int maxRunTime=20);
+	int maxIter=60, int batchSize=1, int pass=20, double a=1, double b=1, double gamma=1,  int maxRunTime=20);
 
 void SVRG_logistic(VectorXd &w, const SparseMatrix<double> &Xt, int* innerIndices, int* outerStarts, const VectorXd &y, double lambda, double eta, VectorXd wtilde, VectorXd G, \
 	int maxIter, int batchSize, int pass, int a, int b, int gamma, const MatrixXd &XtTest, \
